Add 'test_phoenix' Mix env
This way, we also test compilation of the library when the Phoenix
dependency is not available, thus making sure that we guard all
Phoenix-specific code with the `Appsignal.phoenix?` test.

The test directory itself has also been split in a generic and
phoenix-specific directory.

Travis CI runs tests on both the `test` and `test_phoenix` environments.