Don't crash when target platform is not supported
On compile time, a warning is printed when the CPU platform is not
supporte.

On runtime, while loading the NIF, an error is printed. All NIF
stub functions now return stub values and do not call exit() so that
the integration keeps working (without doing anything) when running on
a non-supported OS.