Check valid state on Config.active?
The active function seemed to be missing specs so I added them and added
a check if the config is valid as well. Same behavior as in the Ruby
gem.

https://github.com/appsignal/appsignal-ruby/blob/ca0a92e2c0f140d4cc9a014c3e47ac0582c5a67d/lib/appsignal/config.rb#L114-L116

The valid and active behavior was linked before on the push_api_key
presence, but that was changed in PR #119