Write config to env only when needed (#135)
This allows the config to be initialized multiple times without setting
its options in environment variables. Currently, once the configuration
is written to the environment it its final. Because the environment
variables are leading in the configuration load order (see
http://docs.appsignal.com/elixir/configuration/load-order.html), they
cannot be overwritten once written to the environment.

This change allows us to initialize the config multiple times, without
writing it options to the environment immediately. This is useful for
non-application oriented code such as Mix tasks.