Load agent.ex only once and set on project data
If we load `agent.ex` in the function it gets loaded at runtime, which
doesn't work for released apps.

Instead we load/compile the file first thing when compiling the
AppSignal package. Once it is loaded we set the data on the project's
data and make it accessible to the diagnose task through
`Mix.Project.config[:agent_version]`