Don't add ex_doc to runtime
As recommended by ex_doc in the README:
https://github.com/elixir-lang/ex_doc

What does runtime do?

`:runtime` - whether the dependency is part of runtime applications.
Defaults to true which automatically adds the application to the list of
apps that are started automatically and included in releases

Source: https://hexdocs.pm/mix/Mix.Tasks.Deps.html