Configure agent with underscored config env vars (#151)
* Configure agent with underscored config env vars

During a hot reload the AppSignal process is not reset/cleared/reloaded
in any way to flush the previously set config env vars. When the
AppSignal process starts it writes the current config to the environment
so that the agent can read from this and be initialized with the same
config.

During a hot reload the AppSignal process tries to initialize the config
again, but because the system env still contains the config for the
agent from the last time it initialized the config it reads the previous
config. The env vars are loading in initializing the config, as
described here:
http://docs.appsignal.com/elixir/configuration/load-order.html

The fix? Underscore all the config sent to the agent process so that it
doesn't conflict with the application's config.

* Bump agent to 413c222