Fix unpacking agent tar as root (#179)
When the installation is being run under the root user tar does not use
the `--no-same-owner` option by default. As described in the "tar" docs:

"--no-same-owner extract files as yourself (default for ordinary users)"

Root is not an ordinary user and thus this is not an option.

The problem then is that it tries to apply the same permissions as the
files that were packed, causing errors because those user ids don't
necessarily exist on a user's machine. And if they do it's unlikely that
they are from the user we want.

I was able to reproduce this with a Docker dingy setup as described in
issue #168: https://github.com/appsignal/appsignal-elixir/issues/168