ErrorHandler doesn't handle :error events
Closes #157 & #173.

Before, the ErrorHandler had to match on :error events (besides
:error_report events) to be able to catch errors that were already
caught by Plug.ErrorHandler.

After merging #187, we can drop support for :error events. This drops
support for unsupervised "bare" spawns, and only triggers an error when
a process actually crashes. Also, this removes a lot of code that dealt
with not sending error reports twice (after the :error event _and_ the
:error_report event).