Send diagnostics data report to AppSignal
This change will allow users to send the diagnostics report to
AppSignal. We'll ask before sending if the user wants this, but because
we see how useful this can be when debugging an issue it sends the
report by default. When you press enter, no input, "Y" (yes) is
submitted.

Why this change?
We've seen that in every support request we get we'll ask for the
diagnostics report. To make it easier for everyone this change allows
users to send it directly to AppSignal so our support team can review it
without valuable data getting lost by formatting by our support system
for example.

We also provide the user with a support token when they send the report.
This can be useful for a user if they see a problem with the output and
want to dive right in with the report. They can contact us with the
support token, after which we can immediately review the report.