Transaction.set_request_metadata/2 works without skip_session_data
This patch uses `!` instead of `not` in
Transaction.set_request_metadata/2, which handles non-booleans to be
passed through the skip_session_data configuration.

This caused the flip-flopping error on CI, because the configuration
could be changed in another test. Also, it makes sense to assume nil
means false in this case.