ErrorHandler handles errors in tuples
When an error is triggered from a Task, or another linked process:

    Task.async(fn ->
      raise("Exception")
    end)

The error looks like this:

    {
      %RuntimeError{message: "Exception!"},
      [
        {AppsignalPhoenixExample.PageController, :"-exception/2-fun-0-", 0, [file: 'web/controllers/page_controller.ex', line: 18]},
        {Task.Supervised, :do_apply, 2, [file: 'lib/task/supervised.ex', line: 85]},
        {Task.Supervised, :reply, 5, [file: 'lib/task/supervised.ex', line: 36]},
        {:proc_lib, :init_p_do_apply, 3, [file: 'proc_lib.erl', line: 247]}
      ]
    }

Appsignal.ErrorHandler.extract_reason_and_message/2 currectly returns
"%RuntimeError{message: "Exception!"}" as the name, and includes the
whole stacktrace in the description.

This patch changes the name to "RuntimeError", and the description to
"Process #PID<0.574.0> terminating: Exception!".