1.5-style UndefinedErrors in ErrorExtracterTest
Since [313dc9c] (released in 1.5.0), the message from Elixir's
`Protocol.UndefinedError` lists modules that do implement the missing
protocol:

    Foo error: protocol Enumerable not implemented for {:error, :foo}. This protocol is implemented for: Date.Range, File.Stream, Function, GenEvent.Stream, HashDict, HashSet, IO.Stream, List, Map, MapSet, Range, Stream"

In ErrorExtracterTest, we expected the first part of the message (like
it was before 1.5), but not the suggested modules. To get the tests to pass on
1.5, we'll use a regular expression that matches the first part of the string,
leaving the second part optional.

[313dc9c]: https://github.com/elixir-lang/elixir/commit/313dc9c85059d2f15cc5eb880ab5282fe3c887b2