Don't register Transactions created in send_error
Before, `send_error` reused the current `Transaction` from the
`TransactionRegistry` if one existed. If not, it'd create one, and store
it in the Registry.

Considering the following controller action:

    def index(conn, _params) do
      Appsignal.send_error(%RuntimeError{}, "Oops 1!")
      Appsignal.send_error(%RuntimeError{}, "Oops 2!")
      render conn, "index.html"
    end

When this action was called, a `Transaction` for the web request was
created by the Plug integration. When the first `send_error` was called,
the existing `Transaction` was fetched from the Registry, and completed.

Since c3cabd2, the `Transaction` was removed from the registry by
replacing its pid with `:removed`. The second call to `send_error` would
then get a `nil` transaction and not send anything at all.

After the call was done, only one `Transaction` made it to AppSignal.

Instead, we want to send three. One for the web request, and one for
each call to `send_error`. To do that, this commit evades the registry
completely in `send_error`, so the existing `Transaction` is left to be
completed by the Phoenix Plug, and both `send_error`s get a fresh
transaction.

This poses a problem, though. Before, since the `Transaction` was
reused, a call to send_error would take over the namespace of the
current action. This patch doesn't take this into account, so all calls
to `send_error` result in an error in the `:background_job` namespace.

In the [Ruby implementation], we default to `:http_request`, which seems
like a sensible solution.

[Ruby implementation]: https://github.com/appsignal/appsignal-ruby/blob/eb421a537b8ebaac0d66fc20f002276e9998b489/lib/appsignal.rb#L267