Use :http_request as the namespace in send_error
Before, send_error would use the current Transaction from the Registry.
In a Phoenix controller, the existing Transaction would already have the
`:http_request` namespace, so it worked automatically for send_error.

Since we create a new Transaction in send_error now, we can't use this
trick. The Ruby integration defaults to `:http_request`, and allows you
to override it in its second argument.

This commit defeaults the namespace for Transactions from send_error to
`:http_request`, but doesn't allow overriding it yet. We'll need to find
a backwards-compatible way to add it, but the send_error function
already has six arguments (where the last is a function).