Add Appsignal.send_error/7
Possible solution to pass custom namespaces (other than `:http_request`)
to `Appsignal.send_error/7`. This adds an optional namespace argument
(which defaults to `:http_request`). Using it requires the user to pass
a bunch of default values to get to the seventh argument.

```
Appsignal.send_error(
  %RuntimeError{message: "Some bad stuff happened"},
  "",
  System.stacktrace(),
  %{},
  nil,
  fn(t) -> t end,
  :background_job
)
```

We'll need to revisit this for 2.0.