Remove mix.lock file from repo
We have a mix.ex file where the dependencies are described. The package
should work with these versions. If it doesn't or specific exceptions
are made per dependency version, another strategy of managing and
testing against these dependency versions should be added, e.g.
https://github.com/thoughtbot/appraisal but for Elixir.

The only reason to commit the mix.lock file is to make sure every
developer uses the same versions locally while developing this package,
but we decided against that.