Remove transaction from registry in send error (#260)
* Remove transaction from registry in send error

Before this change calling send error multiple times in the same process
would only post an error once because the submitted transaction was
still around.

* Don't register Transactions created in send_error

Before, `send_error` reused the current `Transaction` from the
`TransactionRegistry` if one existed. If not, it'd create one, and store
it in the Registry.

Considering the following controller action:

    def index(conn, _params) do
      Appsignal.send_error(%RuntimeError{}, "Oops 1!")
      Appsignal.send_error(%RuntimeError{}, "Oops 2!")
      render conn, "index.html"
    end

When this action was called, a `Transaction` for the web request was
created by the Plug integration. When the first `send_error` was called,
the existing `Transaction` was fetched from the Registry, and completed.

Since c3cabd2, the `Transaction` was removed from the registry by
replacing its pid with `:removed`. The second call to `send_error` would
then get a `nil` transaction and not send anything at all.

After the call was done, only one `Transaction` made it to AppSignal.

Instead, we want to send three. One for the web request, and one for
each call to `send_error`. To do that, this commit evades the registry
completely in `send_error`, so the existing `Transaction` is left to be
completed by the Phoenix Plug, and both `send_error`s get a fresh
transaction.

This poses a problem, though. Before, since the `Transaction` was
reused, a call to send_error would take over the namespace of the
current action. This patch doesn't take this into account, so all calls
to `send_error` result in an error in the `:background_job` namespace.

In the [Ruby implementation], we default to `:http_request`, which seems
like a sensible solution.

[Ruby implementation]: https://github.com/appsignal/appsignal-ruby/blob/eb421a537b8ebaac0d66fc20f002276e9998b489/lib/appsignal.rb#L267

* Remove Transactions from assertions in PhoenixTest

* Move send_error tests to AppsignalTest

* Use :http_request as the namespace in send_error

Before, send_error would use the current Transaction from the Registry.
In a Phoenix controller, the existing Transaction would already have the
`:http_request` namespace, so it worked automatically for send_error.

Since we create a new Transaction in send_error now, we can't use this
trick. The Ruby integration defaults to `:http_request`, and allows you
to override it in its second argument.

This commit defeaults the namespace for Transactions from send_error to
`:http_request`, but doesn't allow overriding it yet. We'll need to find
a backwards-compatible way to add it, but the send_error function
already has six arguments (where the last is a function).

* Add Appsignal.send_error/7

Possible solution to pass custom namespaces (other than `:http_request`)
to `Appsignal.send_error/7`. This adds an optional namespace argument
(which defaults to `:http_request`). Using it requires the user to pass
a bunch of default values to get to the seventh argument.

```
Appsignal.send_error(
  %RuntimeError{message: "Some bad stuff happened"},
  "",
  System.stacktrace(),
  %{},
  nil,
  fn(t) -> t end,
  :background_job
)
```

We'll need to revisit this for 2.0.