Use musl build for older systems
With the AppSignal for Elixir package v1.4.0 release we dropped support
for some older Operating Systems that use a libc version older than
2.12.

Since package v0.11.0 we started using a musl build so the libc version
didn't matter on the user's machine. Before that we used a very old
machine that used the libc version from CentOS 5, so we supported some
very old systems.

To now support older Operating System versions, with a libc older than
2.12, this change adds a check for the older versions and selects the
musl build instead.

This should fix the problem for people upgrading to (future) package
1.4.1 and higher.

Also move the build detection to the System module so we can reuse the
logic and add specs for its behavior.