Refactor build architecture matcher
No longer rely on the partial matches of host triples, but rather only
use the system_architecture value to match 32/64-bit systems.

We've seen that the value returned from
`:erlang.system_info(:system_architecture)` is set manually on
installation of Elixir, so relying on it too much could introduce
problems later on when an installation reports the wrong architecture
and uses the default value `x86[_64]-unknown-linux-gnu`.

Match the platform for which to install the agent by using the same
logic used by the AppSignal Ruby gem with
`Appsignal.System.agent_platform`.