Reuse Appsignal.agent module by unloading it
Previously we stored the value of `Appsignal.Agent.version` on the
project map so we could access it inside the project. We only loaded the
`Appsignal.agent` module in the Mix env, because loading it again in the
project itself added a warning about redefining the module.

With this change we unload the module in the Mix env once we're done
installing and load it again in the library context so we can directly
call upon the Appsignal.Agent module functions.