Fetch request_id in Appsignal.Plug.extract_meta_data/1
Since the request_id hasn't been set when the Transaction gets started
in `Appsignal.Plug.call/2`, the previous implementation never actually
found the current request's ID.

This patch generates a Transaction ID and adds the reuqest ID to the
metadata later when extracting the metadata from the conn. This causes
the Transaction ID to be different than the request ID in the metadata,
so we don't rely on the processor falling back to the transaction ID
when no request ID is set.

Closes #282.