Fix improper list data encoder handling
"Fix" this by saving the improper list as a serialized string. We can't
save the real value of the improper list in our extension data encoder
as it's a Erlang specific type.

The prefix of "improper_list" will be parsed on our front-end as an
improper list and displayed as such. Only for data handling it will be
stored as such. This same fix was applied for the "bigint" value in
8b5de8da11e9473c4a9146fb3893f3a2df16df4a.

```
[1, 2, 3 | 4]
"improper_list:[1, 2, 3 | 4]"
```