Allow OTP 18.3 failures
Currently the only OTP version it randomly fails on to test the agent in
diagnose mode. Outside the test env this seems to work fine. Given
enough retries or luck it will eventually work on TravisCI, but it's not
helpful to developers continuously seeing a red mark for build status.