:info instead of :warn when AppSignal is disabled
Closes #335. Thanks for reporting this, @axelson!

If AppSignal is not active, we'll :info-log the "AppSignal disabled"
line. This will show in a Phoenix app in development, but not in tests
(as the default :logger level for tests is :warn in Phoenix).

We'll only warn if the configuration is invalid, but AppSignal was set
to active (like when omitting the push_api_key from
config/appsignal.exs).