Set default request_headers config in installer
This suppresses the warning message about the missing `request_headers`
config option. It's set to an empty list as the config option doesn't
apply to the demo samples that are send later in the installer.
This creates the most minimal config needed in the installer. No need to
set the suggested request_headers here.