Explicit header whitelist in configuration (#336)
* Add request headers to config from install task

* Use 2D array for request headers

* Add APPSIGNAL_REQUEST_HEADERS to appsignal.install

* Remove ip-related fields from default header list

https://github.com/appsignal/appsignal-server/issues/2791#issuecomment-377479183

* Add :request_headers warning to Config.initialize()

* Remove user agent and referer headers from default

Too much personal data being sent that's not GDPR compliant.

* Fix default config headers alignment

* Set default request_headers config in installer

This suppresses the warning message about the missing `request_headers`
config option. It's set to an empty list as the config option doesn't
apply to the demo samples that are send later in the installer.
This creates the most minimal config needed in the installer. No need to
set the suggested request_headers here.