Overwrite message for Phoenix.ActionClauseError
Phoenix.ActionClauseError includes parameters in its exception message.
Although we won't send it from `AppSignal.Plug`, we'd still like to
filter them out when a customer decides to send them to us anyway. We
replace the whole message instead of filtering out the map, which might
break in future versions.