Add appsignal_to_json implementation
Available as `to_map` on the `Appsignal.Transaction` module. Allows a
view of the transaction object as is currently stored in the extension.

Used for testing purposes only, do not use this function in your own
apps!

To fetch transaction data from the transaction the extension needs to be
started with a valid config, otherwise it returns an empty string. Since
we want to do this more often, start the AppSignal app with a valid
config by default. Use the `with_config` helper function if you want to
test an invalid config.

Implement one test, setting an action name on the transaction, to make
sure it works from end to end.

Co-authored-by: Jeff Kreeftmeijer <jeffkreeftmeijer@gmail.com>