Ensure ca_file_path is written to agent env (#381)
PR #380 fixed the ca_file_path option for the integration, but not the
agent config. The default path was not written to the config, only the
user config.

This fix ensures that the ca_file_path config option is part of the
config hash. It now uses the same logic as the rest of the config.

Add more tests to ensure the logic we expect is working.