Link gcc to extension .so library
This fixes the missing `libgcc_s.so` error on AppSignal boot on musl
builds with a two stage deploy.

```
Error loading shared library libgcc_s.so.1: No such file or directory
(needed by /app/lib/appsignal-1.8.1-beta.2/priv/appsignal_extension.so)
```

This was caused by the 1.8 build having more dynamically linked
dependencies than the previous build. This was `libgcc_s`. Since it's
only available as a shared object file (`.so`) we can't specify
`-static` as an option to gcc. Instead, we need to add the
`-static-libgcc` option. Docs:
http://gcc.gnu.org/onlinedocs/gcc/Link-Options.html (search for
`-static-libgcc`).

TODO: This option may need to be put behind a "musl-only" flag, if it's
not necessary or breaks stuff for GNU builds.

Fixes:
https://docs.appsignal.com/support/known-issues/compilation-dependencies-required-at-runtime.html

For more information (private issue):
https://github.com/appsignal/appsignal-agent/issues/375