Fetch and store separate config sources
Add functions to `Appsignal.Config` to fetch the config options from
every source separately using similar functions. Previously we all
merged the config options in the Application config, which made it
impossible to trace back where what value came from for the config
option.

This sources hash will be included in the diagnose report in a future
commit.

On initialize the different functions are read and stored in the sources
map before they're merged together for the final config.
The new values in the map stored in the Application config contain the
config options set for all the possible sources:

- `default`
  returns the default config and runtime config (things we can't
  hardcode in the `Appsignal.Config` default_config module attribute).
- `system`
  options detected from the system host, e.g. Heroku, or if the
  `APPSIGNAL_PUSH_API_KEY` environment variable is set.
- `file`
  options set by the user in the Mix config files, `config.exs` or
  `appsignal.exs`, in the `config/` directory.
- `env`
  options set by the user in the host's system environment using
  environment variables.