Add config option sources to diagnose report
This change print config option sources in diagnose report.
This allows users to track back on their own, based on the output, where
the config options are coming from.
Helps with debugging where a particular option gets set.

It also adds the config options sources to the config report as
described in the format in the PR for the Ruby gem:
https://github.com/appsignal/appsignal-ruby/pull/444

New format:

```
Configuration
  # Option with only the default source
  diagnose_endpoint: https://appsignal.com/diag

  # Option with one source, not default
  name: appsignal-plug-example (Loaded from file)

  # Option with multiple sources
  active: true
    Sources:
      default: false
      file: true
```