Fail AppSignal extension installation on warnings
When a warning occurs during the extension installation, fail the
installation. Currently warnings such as linking issues are not cause
for a failed installation, such as #406.

This was replicated by trying to install the linux-gnu build on macOS,
this did not caused the installation to be registered as "failed". This
was unexpected behavior.

Since we don't think the `install.log` will catch any more problems than
the `install.report` now that it also fails on warnings, remove the
`install.log` and instead print the error in the app's STDOUT so the
user can see if anything is wrong.