Rescue `make` command errors
When `make` is not found `System.cmd` will raise an error, causing the
installation to crash.

```
** (ErlangError) Erlang error: :enoent
    (elixir) lib/system.ex:632: System.cmd("make", [], [stderr_to_stdout: true])
    mix_helpers.exs:173: Mix.Appsignal.Helper.compile/1
    (mix) lib/mix/task.ex:316: Mix.Task.run_task/3
    (mix) lib/mix/tasks/compile.all.ex:68: Mix.Tasks.Compile.All.run_compiler/2
    (mix) lib/mix/tasks/compile.all.ex:52: Mix.Tasks.Compile.All.do_compile/4
    (mix) lib/mix/tasks/compile.all.ex:23: anonymous fn/1 in Mix.Tasks.Compile.All.run/1
    (mix) lib/mix/tasks/compile.all.ex:39: Mix.Tasks.Compile.All.with_logger_app/1
    (mix) lib/mix/task.ex:316: Mix.Task.run_task/3
```

We intend to fail the installation without crashing the compilation step
of the entire app.

Rescue the errors and record a failure of the installation without
actually crashing the installation so the rest of the app can compile.