Add Minutly probing system
Every minute we want to run registered probes to gather metrics from the
Elixir/Erlang app. The system is designed in such a way that we always
run the probe at the full minute. When we hit the full minute we go over
a list of registered probes and create tasks to execute them.

Registering a probe can be done either by passing in a function, or
creating a module:

Module:
```
defmodule ExampleProbe do
  def call do
  # .. Do work
  end
end

Appsignal.Probes.register(:example_probe, &ExampleProbe.call/0)
```

Passing in a function:
```
Appsignal.Probes.register(:example_probe, fn -> SomeWork end)
```

Thing to note, in the tests we run the probing system every 10ms, this
is to keep the tests at a normal speed.

If you want to register default probes that we ship along with the
package you can do so in lib/appsignal.ex -> add_default_probes