Make sure events are added to transactions started with decorators (#468)
f319d28d9f4b3f6cd427aae973c7b6caa1d4886b introduced a regression where
functions decorated with the `transaction` decorator were run before
their transaction was started. This caused any nested events to be
dropped and results in empty transactions being reported.

This was caused by unquoting the body in the `transaction/3` without
wrapping it in an anonymous function, causing the function to run
immediately and its return value being added as the body parameter in
the `in_transaction/3` function.

This patch adds a regression test to make a function with the
`transaction` decorator that calls a function with the
`transaction_event` decorator produces a transaction with an event, and
wraps the function body in an anonymous function that's called in the
`in_transaction/3` function after starting and before finishing the
transaction.