Update make clean task
Update it to clean all files that are in the `priv/` dir of the package.

Also set a default value to the `LIB_DIR` if it's not set. This is
usually set by the `mix_helpers.exs` file instead, but in case we run
`make clean` directoy, it is not set.