Fix musl detection on installation
The command `ldd --version` always returns exit code `1`.
This is because `--version` is not a supported option, it always prints
the version number when it runs into an issue when starting.
It's mostly there so we know what it's doing.

Because it always returns exit code `1`, don't listen to the exit code.
Only check if the command does not exist. That raises an error which we
catch instead and only then return an error tuple.

This behavior broke in #487