Handle non-maps in Config.active?/0 (#495)
We started explicitly checking if the integration was configured before
adding or fetching a Transaction from the Registry in
https://github.com/appsignal/appsignal-elixir/pull/480/files, which was
released in 1.10.5.

In the case described in #494, the configuration isn’t loaded. The call
to `Config.active?/0`  so the call to check the configuration failed.
This patch makes sure to catch that case, so the function returns
`false` if the configuration is anything but a map with both its
`:active` and `:valid` keys containing true.