Add support for both Jason and Poison (#506)
* Switch from Poison to Jason

* Add Json, and json_encoder/0 to mix_helpers

* Switch between Jason and Poison in Json and json_encoder/0

* Add Poison and Jason as optional dependencies

* Automatically switch between Jason and Poison in mix_helpers.exs

* Move JSON encoder error, add encode_report_file/1

Instead of both encoding and writing the report in the
write_report_file/2 function, this patch splits the encoding part to a
new function named encode_report_file/1, which takes a report map and
returns either an :ok- or :error-tuple.

The function checks if there's a JSON encoder available through the
json_encoder/0 function. If there is, the report is passed to the
encoder's encode/1 function, and its result is returned. If there is no
encoder available, the json_encoder/0 function returns an :error-tuple,
which is immediately returned by the encode_report_file/1 function.

By always returning either an :ok- or :error-tuple (and the :error tuple
coming with a reason), the write_report_file/2 function can handle
encoding errors with a single clause.

* Rename Json to Appsignal.Json

* Add MissingEncoder and MissingEncoderError