Move finish_with_conn/2 to before_send
Instead of finishing the transaction in an else-clause of the try-block
that handles errors, finish the transaction in a before_send, which is
always called, even when running into an error. By finishing the
transaction there, we're sure the action name has been resolved,
allowing us to set the metadata here instead of in the Phoenix
instrumenter.