Use Plug.ErrorHandler and Plug.Test in Appsignal.PlugTest
To more closely resemble an actual Plug app, the Plug tests now use
Plug's test helpers and Plug.ErrorHandler instead of implementing
`call/2` manually.

NOTE: The Plug.BadRequestError test is disabled, as it has a bug that
causes some errors to be reported when they shouldn't, which will be
fixed shortly.