Add Phoenix Telemetry event handler to add Phoenix 1.5 instrumentation (#528)
* Use Plug.ErrorHandler and Plug.Test in Appsignal.PlugTest

To more closely resemble an actual Plug app, the Plug tests now use
Plug's test helpers and Plug.ErrorHandler instead of implementing
`call/2` manually.

NOTE: The Plug.BadRequestError test is disabled, as it has a bug that
causes some errors to be reported when they shouldn't, which will be
fixed shortly.

* Unwrap Plug.Conn.WapperErrors before testing for plug_status

In some situations, errors were reported with plug_status under 500
because they were wrapped in a Plug.Conn.WrapperError. This patch makes
sure to test the plug_status right before passing the error on to the
ErrorHandler.

* Use ErrorHandler.set_error/3 from Appsignal.Plug.do_handle_error/3

* Move finish_with_conn/2 to before_send

Instead of finishing the transaction in an else-clause of the try-block
that handles errors, finish the transaction in a before_send, which is
always called, even when running into an error. By finishing the
transaction there, we're sure the action name has been resolved,
allowing us to set the metadata here instead of in the Phoenix
instrumenter.

* Move call to try_set_action/2 to finish_with_conn/2

* Don't set the action name from the Phoenix instrumenter

* Add Phoenix.EventHandler, start endpoint events

* Add stop endpoint events in Phoenix.EventHandler

* Add Telemetry to dependencies

* Automatically attach Appsignal.Phoenix.EventHandler

* Move transaction starting to callback function in Phoenix.EventHandlerTest

* Phoenix.EventHandler uses with in-conn transactions, if available

* Add typespecs to Phoenix.EventHandler

* Use empty maps instead of nils as transaction bodies