Set Phoenix 1.5 action names in the EventHandler (#538)
* Use FakeTransaction.started_events/1 to test started events in instrumenter_test

* Set Phoenix action names in the before_send callback

From Phoenix 1.5 on, we lose the Phoenix.Instrumenter we relied on
previously. Instead, we'll now use the telemetry events that are emitted
by the Phoenix framework to create an event for the whole endpoint.

Since we don't have access to anything inside the event like we did when
we had the Phoenix.Instrumenter, we can't set the action name before the
action occurs, so we need to use a before_send callback to set the
action name after the action is finished, but before a response is sent.

Before, we set the action name by calling the set_action function on the
Nif. Users could overwrite the action name in their action by calling
that same function. Since we set the action name after the action is
complete now, we need to make sure we don't overwrite what's already
been set.

To fix that, we now allow setting an action name on the conn. If an
action name is set in there, we won't overwrite it at the end of the
request.

```
def index(conn, _params) do
  conn
  |> Appsignal.Plug.set_action("OVERWRITTEN")
  |> render("index.html")
end
```

* Set action name in Phoenix.EventHandler

* Restore else-clause to finish transactions in Appsignal.Plug

* Set action name before call in Appsignal.Plug