Use `__STACKTRACE__ /0` on Elixir >= 1.7 (#559)
* Add Appsignal.Stacktrace.get/0 to wrap System.stacktrace/0

* Add system-stacktrace to CI branches

* Use :erlang.get_stacktrace/0 from Appsignal.Stacktrace.get/0

* Reorder aliases in Channel and LiveView

* fixup! Add Appsignal.Stacktrace.get/0 to wrap System.stacktrace/0

* Use __STACKTRACE__ for Elixir >= 1.7

Appsignal.Stacktrace checks the Elixir version at compile time and
defines a .get/0 macro that uses __STACKTRACE__ for Elixir versions 1.7
and up, and :erlang.get_stacktrace/0 for all versions before that.

This is a macro, as these can only be called inside a catch clause. The
module has to be required in any module that uses it because of that.

* fixup! Add Appsignal.Stacktrace.get/0 to wrap System.stacktrace/0