Don't register query spans without parents (#600)
As reported via [support](https://app.intercom.com/a/apps/yzor8gyw/inbox/inbox/conversation/16410700017831).

Since 2.0, we're not dropping Spans anymore when they don't have a parent Span.
While that makes technical sense, users are seeing spikes in usage because
previously ignored queries that happened outside of any web requests are now
instrumented.

This patch reverts that behaviour to how it worked in 1.x. If a query happens
when there's no parent to conatain it, the query isn't instrumented.