fix: use get_env/3 to avoid compilation errors (#644)
Fixes bug occurring when no config was specified for AppSignal. The
`debug?` function would try to load the config even if no config was
loaded, which caused an error trying to fetch a key from an undefined
config.

```
** (ArgumentError) could not fetch application environment :config for
  application :appsignal because configuration at :config was not set
```

This issue was introduced by PR #642 which added a check to the logger
to see if the AppSignal config was set to `debug` is `true`. And would
occur in issue #643 when called by the plug and phoenix packages:

- https://github.com/appsignal/appsignal-elixir-phoenix/pull/12
- https://github.com/appsignal/appsignal-elixir-plug/pull/11

Fixes #643

An alternative solution was considered, to fall back on an empty list
instead, but this solution was chosen as it uses the `@default_config`
and thus replicates the default behavior as when the config was loaded
but not altered.

```
# Example alternative solutino
Application.get_env(:appsignal, :config, [])[:debug] || false
```

---

Original author of the change is @the-mikedavis.
Commit message updated before merge by @tombruijn to include all
information from issue #643 and debugging for future reference.