Update APPSIGNAL_BUILD_FOR_MUSL flag behavior
Currently the APPSIGNAL_BUILD_FOR_MUSL flag listens to any value and
accepts that as meaning that the user intends to install the Linux musl
build.

The defined behavior in the integrations guide specifies that the
APPSIGNAL_BUILD_FOR_MUSL env var should listen to the `1` or `true`
values.

Source:
https://github.com/appsignal/integration-guide/blob/9b6fb1491e3aab87cb29b732470c8162f9cd080b/build/installation.md#appsignal_build_for_musl

Update the APPSIGNAL_BUILD_FOR_MUSL flag to listen to only the `1` and
`true` values as agreed upon.