Add mono to project
Use our own tooling for development in the Elixir integration repos.
https://github.com/appsignal/mono/
Follow the instructions on the mono README to install it locally.

Mono is also installed on the CI using the `script/setup` script
which is run before every job. The mono download is cached between
jobs so we don't download it for every job fresh. Update the version
variable `MONO_VERSION` to update the mono install on the CI.
This `setup` script will check the existence of the `cache` command, if
not it will skip it. This allows it to be reused in the test-setups
repo.

Mono repo support for Elixir isn't 100% feature complete yet so we're
first adding it to every repo separately.

I didn't add mono to the Semaphore CI test setup as it doesn't really
add anything there at this point. I also didn't add it to the Ruby gem
CI yet either.