Add Linux ARM override value to diagnose report
In PR #656 I added the experimental Linux ARM build behind a feature
flag. In the implementation I forgot to add the `linux_arm_override`
value to the diagnose report with which we can detect if this env var
was set during installation. This change adds that value to the report.