Transaction and channel action decorators only create root spans (#661)
* Use "http_request" namespace for test spans

Previously, the namespaces used in the tracer_test indicated the span's
type (using namespaces like `"root"` and `"child"`), as they were
unused.

To clarify that these are namespaces, and to allow future changes where
the namespaces might become significant whwne creating spans, the tests
now all use `"http_request"` as the namespace.

* Add regression test for child span namespaces

To make sure we create a child span even if there's a namespace
mismatch. The namespace passed in the child span is ignored, and the new
span is added to the current one as you can't have multiple traces in
one process.

* Transaction decorator only creates root spans

Currently, because the `transaction` decorator uses the
`Appsignal.Instrumentation.Decorators.instrument/2` function, the
decorator creates child spans when one already exist.

The `transaction` decorator exists for backward compatibility, and
originally only created Transactions, which are now root spans.

This patch makes sure the decorator always creates a root span, even if
another span already exists.

* Channel action decorator only creates root spans

Like the `transaction` decorator, the `channel_action` decorator should
only create root spans.

* Add changeset

> "Only create root spans from transaction and channel action
> decorators, as they're meant to only be used when no span exists yet."

* Add regression test for root span order

This patch adds a regression test to prove root spans are created and
closed in order.

* Add regression tests for instrument_root/3

This patch adds regression tests for instrument_root/3, to make sure
they actually create root spans through the tracer by passing an
explicit nil as the parent span.

* Don't set namespace twice in instrument_root

Instead of setting the namespace when creating the span and then
overwriting it using `set_namespace`, pass the namespace immediately
when starting the span.

* Accept atoms in transaction/3

To restore the original implementation of the `transaction` decorator,
this patch adds a clause that converts atom namespaces to strings before
calling out to `Appsignal.Instrumentation.instrument_root/3`.