Fix Apple ARM detection
According to support issue #150
https://github.com/appsignal/support/issues/150 the String reported by
Elixir as the system architecture is `arm-apple-darwin20.2.0` by
`:erlang.system_info(:system_architecture)`.

There are two problems with that String:

The first issue is that the `arm` String was not accounted for, only
the `aarch64` String. By adding the `arm` String we can also match ARM
architectures that report themselves as `arm`.

Secondly, it uses `arm` and not `arm64`, so we don't know which
architecture it could be. This is more of a problem for other
(non-darwin) hosts that may have 32-bit architecture. I decided not to
account for this, I don't know how, and it will fail later on in the
installation process with an architecture mismatch.