Replace decorator slashes with underscores
Because the Span API uses a Span's name as what was previously described
as the event name, underscores are no longer supported.

Although this was the (documented) case before, changes in the
processing API caused this to break for people using the decorators
(which are a fallback for users upgrading from 1.0).

This patch changes the slashes in automatically generated span names to
underscores, as that's the only available alternative.

It's marked as a minor patch in the changelog, as this changes some
reported names, causing newly created incidents for insidents that
already existed before.

https://docs.appsignal.com/api/event-names.html#event-namin

Closes https://github.com/appsignal/support/issues/132