Remove dot prefix from bin/setup script in CI
The dot is not necessary to run the `bin/setup` script, remove it. With
the dot present the scripts will not fail the build if they exit with a
failure. So removing them actually highlights which jobs failed. It
only appears to be a problem for the "mix compile --warnings-as-errors",
"mix format", credo and dialyzer builds for some reason.